+++ src/FetchUtil.ts
@@ -1,6 +1,7 @@
 import SinkMap from '@rdfjs/sink-map'
 import { EventEmitter } from 'events'
 import { Stream } from 'rdf-js'
+import url from 'url'
 import ResponseWrapper from './ResponseWrapper'
 import { merge } from './helpers/MergeHeaders'
 
@@ -10,7 +11,12 @@ function requestAcceptHeaders (sinkMap: Parsers) {
     return [...sinkMap.keys()].join(', ')
 }
 
-async function getResponse (uri: string, method: string, headers: HeadersInit = {}, parsers: Parsers, body?: BodyInit) {
+async function getResponse (uri, { method, headers = {}, body, baseUri, parsers }: { method: string; headers?: HeadersInit; body?: BodyInit; baseUri?: string; parsers: Parsers }) {
+    let effectiveUri = uri
+    if (uri.match(/^https?:\/\//) === null && baseUri) {
+        effectiveUri = url.resolve(baseUri, uri)
+    }
+
     const defaultHeaders: HeadersInit = {
         accept: requestAcceptHeaders(parsers),
     }
@@ -25,20 +31,21 @@ async function getResponse (uri: string, method: string, headers: HeadersInit = 
 
     requestInit.headers = merge(new Headers(defaultHeaders), new Headers(headers))
 
-    const res = await fetch(uri, requestInit)
+    const res = await fetch(effectiveUri, requestInit)
 
-    return new ResponseWrapper(uri, res)
+    return new ResponseWrapper(effectiveUri, res)
 }
 
-export function fetchResource (uri: string, parsers: Parsers, headers?: HeadersInit): Promise<ResponseWrapper> {
-    return getResponse(uri, 'get', headers, parsers)
+export function fetchResource (uri: string, requestInit: { parsers: Parsers; headers?: HeadersInit; baseUri?: string }): Promise<ResponseWrapper> {
+    return getResponse(uri, {
+        method: 'get',
+        ...requestInit,
+    })
 }
 
 export function invokeOperation (
     method: string,
     uri: string,
-    parsers: Parsers,
-    headers?: HeadersInit,
-    body?: BodyInit): Promise<ResponseWrapper> {
-    return getResponse(uri, method, headers, parsers, body)
+    requestInit: { parsers: Parsers; headers?: HeadersInit; body?: BodyInit; baseUri?: string }): Promise<ResponseWrapper> {
+    return getResponse(uri, { method, ...requestInit })
 }