+++ tests/FetchUtil-spec.ts
@@ -30,7 +30,7 @@ describe('FetchUtil', () => {
                 .returns(responseBuilder().body(Bodies.someJsonLd).build())
 
             // when
-            await fetchUtil.fetchResource('http://example.com/resource', parsers, {})
+            await fetchUtil.fetchResource('http://example.com/resource', { parsers })
 
             // then
             const requestHeaders = windowFetch.firstCall.args[1].headers
@@ -44,8 +44,11 @@ describe('FetchUtil', () => {
                 .returns(responseBuilder().body(Bodies.someJsonLd).build())
 
             // when
-            await fetchUtil.fetchResource('http://example.com/resource', parsers, {
-                'x-foo': 'bar',
+            await fetchUtil.fetchResource('http://example.com/resource', {
+                parsers,
+                headers: {
+                    'x-foo': 'bar',
+                },
             })
 
             // then
@@ -61,9 +64,10 @@ describe('FetchUtil', () => {
                 .returns(responseBuilder().body(Bodies.someJsonLd).build())
 
             // when
-            await fetchUtil.fetchResource('http://example.com/resource', parsers, {
-                'x-foo': 'bar',
-            })
+            await fetchUtil.fetchResource('http://example.com/resource', { parsers,
+                headers: {
+                    'x-foo': 'bar',
+                } })
 
             // then
             const requestHeaders = windowFetch.firstCall.args[1].headers
@@ -77,13 +81,27 @@ describe('FetchUtil', () => {
                 .returns(responseBuilder().body(Bodies.someJsonLd).build())
 
             // when
-            await fetchUtil.fetchResource('http://example.com/resource', parsers, {
-                'accept': 'application/vnd.custom+rdf',
-            })
+            await fetchUtil.fetchResource('http://example.com/resource', { parsers,
+                headers: {
+                    'accept': 'application/vnd.custom+rdf',
+                } })
 
             // then
             const requestHeaders = windowFetch.firstCall.args[1].headers
             expect(requestHeaders.get('accept')).toBe('application/vnd.custom+rdf')
+        })
+
+        it('should resolve relative URI against', async () => {
+            // given
+            windowFetch.withArgs('http://example.com/resource')
+                .returns(responseBuilder().body(Bodies.someJsonLd).build())
+
+            // when
+            await fetchUtil.fetchResource('resource', { parsers, baseUri: 'http://example.com/foo/' })
+
+            // then
+            const uri = windowFetch.firstCall.args[0]
+            expect(uri).toEqual('http://example.com/foo/resource')
         })
 
         afterEach(() => {
@@ -98,7 +116,7 @@ describe('FetchUtil', () => {
                 .returns(responseBuilder().body(Bodies.someJsonLd).build())
 
             // when
-            await fetchUtil.invokeOperation('get', 'http://example.com/resource', parsers, {}, 'foo')
+            await fetchUtil.invokeOperation('get', 'http://example.com/resource', { parsers, body: 'foo' })
 
             // then
             const body = windowFetch.firstCall.args[1].body
@@ -111,9 +129,10 @@ describe('FetchUtil', () => {
                 .returns(responseBuilder().body(Bodies.someJsonLd).build())
 
             // when
-            await fetchUtil.invokeOperation('get', 'http://example.com/resource', parsers, {
-                'x-foo': 'bar',
-            }, 'foo')
+            await fetchUtil.invokeOperation('get', 'http://example.com/resource', { parsers,
+                headers: {
+                    'x-foo': 'bar',
+                } })
 
             // then
             const requestHeaders = windowFetch.firstCall.args[1].headers
@@ -128,9 +147,10 @@ describe('FetchUtil', () => {
                 .returns(responseBuilder().body(Bodies.someJsonLd).build())
 
             // when
-            await fetchUtil.invokeOperation('get', 'http://example.com/resource', parsers, {
-                'x-foo': 'bar',
-            }, 'foo')
+            await fetchUtil.invokeOperation('get', 'http://example.com/resource', { parsers,
+                headers: {
+                    'x-foo': 'bar',
+                } })
 
             // then
             const requestHeaders = windowFetch.firstCall.args[1].headers
@@ -144,9 +164,10 @@ describe('FetchUtil', () => {
                 .returns(responseBuilder().body(Bodies.someJsonLd).build())
 
             // when
-            await fetchUtil.invokeOperation('get', 'http://example.com/resource', parsers, {
-                'accept': 'application/vnd.custom+rdf',
-            }, 'foo')
+            await fetchUtil.invokeOperation('get', 'http://example.com/resource', { parsers,
+                headers: {
+                    'accept': 'application/vnd.custom+rdf',
+                } })
 
             // then
             const requestHeaders = windowFetch.firstCall.args[1].headers
@@ -159,11 +180,24 @@ describe('FetchUtil', () => {
                 .returns(responseBuilder().body(Bodies.someJsonLd).build())
 
             // when
-            await fetchUtil.invokeOperation('post', 'http://example.com/resource', parsers, {}, new FormData())
+            await fetchUtil.invokeOperation('post', 'http://example.com/resource', { parsers, body: new FormData() })
 
             // then
             const request = windowFetch.firstCall.args[1]
             expect(request.headers.get('content-type')).toBeNull()
+        })
+
+        it('should resolve relative URI against', async () => {
+            // given
+            windowFetch.withArgs('http://example.com/resource')
+                .returns(responseBuilder().body(Bodies.someJsonLd).build())
+
+            // when
+            await fetchUtil.invokeOperation('get', 'resource', { parsers, body: 'foo', baseUri: 'http://example.com/foo/' })
+
+            // then
+            const uri = windowFetch.firstCall.args[0]
+            expect(uri).toEqual('http://example.com/foo/resource')
         })
 
         afterEach(() => {